Rename Model01-MouseKeys to Keyboardio-MouseKeys
The mousekeys are not specific to the Model01, name it after Keyboardio
instead, as suggested by @obra.
