+++ libraries/Keyboardio-MouseKeys/library.properties
@@ -1,8 +1,8 @@
-name=Model01-MouseKeys
+name=Keyboardio-MouseKeys
 version=0.0.1
 author=Jesse Vincent
 maintainer=Jesse Vincent <jesse@keyboard.io>
-sentence=Mouse keys for the Keyboardio Model 01.
+sentence=Mouse keys for Keyboardio boards.
 paragraph=...
 category=Communication
 url=https://github.com/keyboardio/KeyboardioFirmware
